# swagger.yaml
openapi: 3.0.0
info:
  title: Teamitaka Practice API
  version: 1.0.0
  description: |
    Teamitaka 프로젝트 연습을 위한 API 명세서로서, 사용자 관리, 인증, 프로젝트 관리, 모집 공고, 프로필, 댓글, 스크랩, 검색, 대학 인증 등의 기능을 제공합니다.
    이 API는 Node.js와 Express.js를 기반으로 구현되었으며, Swagger UI를 통해 시각화됩니다.  
    주요 기능:
    - 사용자 등록 및 인증 (JWT 기반)
    - 관리자 기능 (관리자 계정 관리, 인증된 이메일 관리)
    - 모집 공고 생성/수정/삭제 및 댓글 관리
    - 프로젝트 생성/수정/삭제 및 멤버 관리
    - 프로필 생성/수정/삭제 및 이력서 조회
    - 스크랩 기능
    - 키워드 기반 검색
    - 대학 인증 (Univecert API 연동)
  contact:
    name: Teamitaka Support
    email: support@teamitaka.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:{port}
    description: 로컬 개발 서버
    variables:
      port:
        default: "3000"
        description: "서버 포트 (동적으로 설정 가능)"
  - url: https://api.teamitaka.com
    description: 프로덕션 서버
...
paths:
...
  /api/auth/validate-password:
    post:
      tags:
        - Auth
      summary: 비밀번호 검증
      description: "제공된 비밀번호가 유효한지 검증합니다 (예: 비밀번호 정책 준수 여부)."
      operationId: validatePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: "password123"
                  description: 검증할 비밀번호
              required:
                - password
      responses:
        '200':
          description: 비밀번호 검증 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  valid: true
                message: "비밀번호가 유효합니다."
        '400':
          description: 비밀번호 형식 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "비밀번호는 최소 8자 이상이어야 하며, 문자와 숫자를 포함해야 합니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"

  # 관리자 관련 엔드포인트
...
  /api/recruitments/{recruitment_id}/comment:
    get:
      tags:
        - Comment
      summary: 댓글 목록 조회
      description: 특정 모집 공고에 달린 댓글 목록을 페이지네이션으로 반환합니다.
      operationId: getComments
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 댓글을 조회할 모집 공고 ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: 페이지 번호
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: 페이지당 항목 수
        - name: sort
          in: query
          schema:
            type: string
            enum: ["created_at_asc", "created_at_desc"]
            default: "created_at_desc"
          description: 정렬 기준
      responses:
        '200':
          description: 댓글 목록 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  comments:
                    - comment_id: 1
                      content: "좋은 공고네요!"
                      user_id: 1
                      username: "johndoe"
                      recruitment_id: 1
                      created_at: "2025-02-27T11:00:00.000Z"
                      updated_at: "2025-02-27T11:00:00.000Z"
                  total: 1
                  page: 1
                  limit: 10
                message: ""
        '404':
          description: 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components.schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
    post:
      tags:
        - Comment
      summary: 댓글 작성
      description: 특정 모집 공고에 댓글을 작성합니다. 욕설 필터링이 적용됩니다.
      operationId: createComment
      security:
        - bearerAuth: []
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 댓글을 작성할 모집 공고 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 500
                  example: "좋은 공고네요!"
                  description: 댓글 내용 (1~500자, 욕설 금지)
              required:
                - content
      responses:
        '201':
          description: 댓글 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  comment_id: 1
                  content: "좋은 공고네요!"
                  user_id: 1
                  recruitment_id: 1
                  created_at: "2025-02-27T11:00:00.000Z"
                message: "댓글 작성 완료"
        '400':
          description: 필수 필드 누락 또는 욕설 포함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "댓글에 부적절한 언어가 포함되었습니다."
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '404':
          description: 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "모집 공고를 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"
  /api/recruitments/{recruitment_id}/comment/{comment_id}:
    put:
      tags:
        - Comment
      summary: 댓글 수정
      description: 특정 모집 공고의 특정 댓글을 수정합니다. 욕설 필터링이 적용됩니다.
      operationId: updateComment
      security:
        - bearerAuth: []
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 댓글이 속한 모집 공고 ID
        - name: comment_id
          in: path
          required: true
          schema:
            type: integer
          description: 수정할 댓글 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 500
                  example: "수정된 댓글입니다!"
                  description: 수정할 댓글 내용 (1~500자, 욕설 금지)
              required:
                - content
      responses:
        '200':
          description: 댓글 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  comment_id: 1
                  content: "수정된 댓글입니다!"
                  user_id: 1
                  recruitment_id: 1
                  updated_at: "2025-02-27T11:00:00.000Z"
                message: "댓글 수정 완료"
        '400':
          description: 필수 필드 누락 또는 욕설 포함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "댓글에 부적절한 언어가 포함되었습니다."
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "댓글 수정 권한이 없습니다."
        '404':
          description: 댓글 또는 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "댓글을 찾을 수 없습니다."
        '500':
          example:
            success: false
            message: "서버 오류 발생"
    delete:
      tags:
        - Comment
      summary: 댓글 삭제
      description: 특정 모집 공고의 특정 댓글을 삭제합니다.
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: recruitment_id
          in: path
          required: true
          schema:
            type: integer
          description: 댓글이 속한 모집 공고 ID
        - name: comment_id
          in: path
          required: true
          schema:
            type: integer
          description: 삭제할 댓글 ID
      responses:
        '200':
          description: 댓글 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data: {}
                message: "댓글 삭제 완료"
        '401':
          description: 인증되지 않은 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "인증되지 않았습니다."
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "댓글 삭제 권한이 없습니다."
        '404':
          description: 댓글 또는 모집 공고 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "댓글을 찾을 수 없습니다."
        '500':
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: '#/components.schemas/ErrorResponse'
              example:
                success: false
                message: "서버 오류 발생"

  # 스크랩 관련 엔드포인트
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        data:
          type: object
          description: 반환 데이터
        message:
          type: string
          description: 추가 메시지
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          description: 요청 실패 여부
        message:
          type: string
          description: 오류 메시지
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 기반 인증 토큰 (Bearer 형식)
...